Move annotation::{Actions,ActionType} to actions
